Cleanup handling of clamps / NaN filters

+++ Source/astcenc_compress_symbolic.cpp
@@ -1235,7 +1235,7 @@ void compress_block(
 	float lowest_correl;
 	float best_errorval_in_mode;
 
-	int start_trial = bsd->texel_count < TUNE_MAX_TEXELS_MODE0_FASTPATH ? 1 : 0;
+	int start_trial = bsd->texel_count < (int)TUNE_MAX_TEXELS_MODE0_FASTPATH ? 1 : 0;
 	for (int i = start_trial; i < 2; i++)
 	{
 		compress_symbolic_block_fixed_partition_1_plane(
